# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indian Railways LocoNet Fault Monitoring System API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Fault:
      type: object
      properties:
        id:
          type: integer
        uniqueFaultId:
          type: string
        alertType:
          type: string
          nullable: true
        locoNo:
          type: string
          nullable: true
        locoType:
          type: string
          nullable: true
        shed:
          type: string
          nullable: true
        zone:
          type: string
          nullable: true
        loggedTimestamp:
          type: string
          nullable: true
        faultCode:
          type: string
          nullable: true
        faultDescription:
          type: string
          nullable: true
        basicUnit:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        moduleName:
          type: string
          nullable: true
        coachNumber:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        datapack:
          type: string
          nullable: true
        coachType:
          type: string
          nullable: true
        recoveryStatus:
          type: string
          nullable: true
        downloadStatus:
          type: string
          nullable: true
        apiTimestamp:
          type: string
          nullable: true
        dbTimestamp:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - uniqueFaultId
        - createdAt
    FaultListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Fault"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - pageSize
        - totalPages
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
paths:
...
  /faults:
    get:
      operationId: listFaults
      tags: [faults]
      summary: List faults with filters and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: zone
          in: query
          schema:
            type: string
        - name: shed
          in: query
          schema:
            type: string
        - name: locoNo
          in: query
          schema:
            type: string
        - name: coachNumber
          in: query
          schema:
            type: string
        - name: faultCode
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: moduleName
          in: query
          schema:
            type: string
        - name: basicUnit
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: coachType
          in: query
          schema:
            type: string
        - name: recoveryStatus
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Faults list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaultListResponse"
  /faults/{id}:
    get:
      operationId: getFault
      tags: [faults]
      summary: Get single fault
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fault record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fault"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /faults/{id}/history:
    get:
      operationId: getFaultHistory
      tags: [faults]
      summary: Get fault history (all occurrences of same fault code + loco)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fault history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fault"
  /faults/export:
    get:
      operationId: exportFaults
      tags: [faults]
      summary: Export faults (returns download token)
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, excel, pdf]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: zone
          in: query
          schema:
            type: string
        - name: shed
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: recoveryStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Export info
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadUrl:
                    type: string
                  filename:
                    type: string
                  records:
                    type: integer
                required:
                  - downloadUrl
                  - filename
                  - records
  /faults/filter-options:
    get:
      operationId: getFaultFilterOptions
      tags: [faults]
      summary: Get distinct values for filter dropdowns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Filter options
          content:
            application/json:
              schema:
                type: object
                properties:
                  zones:
                    type: array
                    items:
                      type: string
                  sheds:
                    type: array
                    items:
                      type: string
                  categories:
                    type: array
                    items:
                      type: string
                  faultCodes:
                    type: array
                    items:
                      type: string
                  modules:
                    type: array
                    items:
                      type: string
                  basicUnits:
                    type: array
                    items:
                      type: string
                  locations:
                    type: array
                    items:
                      type: string
                  coachTypes:
                    type: array
                    items:
                      type: string
                required:
                  - zones
                  - sheds
                  - categories
                  - faultCodes
                  - modules
                  - basicUnits
                  - locations
                  - coachTypes
...